reorganizing components
