+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import './globals.css';
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
-import { ThemeProvider } from '@/components/theme-provider';
+import { ThemeProvider } from '@/components/header/theme-provider';
 
 const inter = Inter({ subsets: ['latin'] });
