+++ components/chat/CopyButton.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import { Copy } from "lucide-react"
-import { Button } from "./ui/button"
+import { Button } from "../ui/button"
 import { useState } from "react"
 
 export function CopyButton() {
@@ -21,7 +21,7 @@ export function CopyButton() {
     <Button 
       variant="ghost" 
       size="icon" 
-      className="absolute -top-1 -right-2 h-8 w-8 hover:bg-muted/50 z-10 rounded-md opacity-70 hover:opacity-100 transition-opacity"
+      className="absolute -top-1 -right-2 h-8 w-8 hover:bg-muted/50 rounded-md opacity-70 hover:opacity-100 transition-opacity"
       onClick={copy}
     >
       <Copy className={`h-4 w-4 ${copied ? 'text-green-500' : ''}`} />
